\name{kid}
\alias{kid}
\docType{data}
\title{
Dataset from Cowling et al. 
}
\description{
Influenza immune response biomarkers dataset. 
}
\usage{data("kid")}
\format{
  A data frame with 736 observations on the following 10 variables.
  \describe{
    \item{\code{hhID}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{intervention}}{a character vector}
    \item{\code{vaccine}}{a numeric vector}
    \item{\code{vaccine.date}}{a Date}
    \item{\code{postvax.date}}{a Date}
    \item{\code{prevax.B.Brisbane}}{a numeric vector}
    \item{\code{postvax.B.Brisbane}}{a numeric vector}
    \item{\code{surv_time}}{a numeric vector}
    \item{\code{flu}}{a numeric vector}
  }
}
\references{
Cowling, B. J., Lim, W. W., Perera, R. A., Fang, V. J., Leung, G. M., Peiris, J. M., & Tchetgen Tchetgen, E. J. (2019). Influenza hemagglutination-inhibition antibody titer as a mediator of vaccine-induced protection for influenza B. Clinical Infectious Diseases, 68(10), 1713-1717.
}
