\name{kyotil}
\alias{kyotil}
\title{kyotil}
\description{
Utility functions by Youyi Fong and Krisz Sebestyen, and some functions copied from other packages for convenience (acknowledged on their manual pages).

Most useful functions: mypostscript/mypdf, mytex, %+% (concatenate strings, e.g. "a" %+% "b"), getFormattedSummary. 

See the Index link below for a list of available functions.
}
